feat(login): refactor code
- fix travis error
- make tests more readable

[Delivers #162948803]
