+++ .travis.yml
@@ -17,7 +17,7 @@ before_script:
   - ./cc-test-reporter before-build
 
 script:
-  - npm test
+  - npm test -- -u
   - npm run build
 
 after_success:
