+++ src/components/InputBox/InputBox.test.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import InputBox from './';
+import InputBox from '.';
 
 describe('InputBox', () => {
+
   it('should be defined', () => {
     expect(InputBox).toBeDefined();
   });
+
   it('should render correctly', () => {
     const tree = shallow(
       <InputBox
